存在问题：记账的时候无法获取当前登录的用户

+++ src/main/java/com/example/jianzhang/dto/record/CreateOrUpdateRecordDTO.java
@@ -19,7 +19,7 @@ import java.util.Date;
 @NoArgsConstructor
 public class CreateOrUpdateRecordDTO {
     @NotNull(message = "{record.spendCategory.not-empty}")
-    private Integer spendCategory;
+    private Long spendCategory;
 
     @NotNull(message = "{record.amount.not-empty}")
     private Double amount;
